updates local fork

+++ code/game/machinery/kitchen/smartfridge.dm
@@ -1,3 +1,4 @@
+
 /* SmartFridge.  Much todo
 */
 /obj/machinery/smartfridge
@@ -15,7 +16,8 @@
 	var/icon_on = "smartfridge"
 	var/icon_off = "smartfridge-off"
 	var/icon_panel = "smartfridge-panel"
-	var/item_quants = list()
+	var/list/item_records = list()
+	var/datum/stored_items/currently_vending = null	//What we're putting out of the machine.
 	var/seconds_electrified = 0;
 	var/shoot_inventory = 0
 	var/locked = 0
@@ -36,6 +38,9 @@
 /obj/machinery/smartfridge/Destroy()
 	qdel(wires)
 	wires = null
+	for(var/datum/stored_items/S in item_records)
+		qdel(S)
+	item_records = null
 	return ..()
 
 /obj/machinery/smartfridge/proc/accept_check(var/obj/item/O as obj)
@@ -153,21 +158,19 @@
 			overlays += "drying_rack_drying"
 
 /obj/machinery/smartfridge/drying_rack/proc/dry()
-	for(var/obj/item/weapon/reagent_containers/food/snacks/S in contents)
-		if(S.dry) continue
-		if(S.dried_type == S.type)
-			S.dry = 1
-			item_quants[S.name]--
-			S.name = "dried [S.name]"
-			S.color = "#AAAAAA"
-			S.loc = loc
-		else
-			var/D = S.dried_type
-			new D(loc)
-			item_quants[S.name]--
-			qdel(S)
-		return
-	return
+	for(var/datum/stored_items/I in item_records)
+		for(var/obj/item/weapon/reagent_containers/food/snacks/S in I.instances)
+			if(S.dry || !I.get_specific_product(get_turf(src), S)) continue
+			if(S.dried_type == S.type)
+				S.dry = 1
+				S.name = "dried [S.name]"
+				S.color = "#AAAAAA"
+				stock_item(S)
+			else
+				var/D = S.dried_type
+				new D(get_turf(src))
+				qdel(S)
+			return
 
 /obj/machinery/smartfridge/process()
 	if(stat & (BROKEN|NOPOWER))
@@ -207,46 +210,27 @@
 		return
 
 	if(accept_check(O))
-		if(contents.len >= max_n_of_items)
-			user << "<span class='notice'>\The [src] is full.</span>"
-			return 1
-		else
-			user.remove_from_mob(O)
-			O.loc = src
-			if(item_quants[O.name])
-				item_quants[O.name]++
-			else
-				item_quants[O.name] = 1
-			user.visible_message("<span class='notice'>[user] has added \the [O] to \the [src].</span>", "<span class='notice'>You add \the [O] to \the [src].</span>")
+		if(!user.remove_from_mob(O))
+			return
+		stock_item(O)
+		user.visible_message("<span class='notice'>\The [user] has added \the [O] to \the [src].</span>", "<span class='notice'>You add \the [O] to \the [src].</span>")
 
-			nanomanager.update_uis(src)
-
-	else if(istype(O, /obj/item/weapon/storage/bag))
+	else if(istype(O, /obj/item/weapon/storage))
 		var/obj/item/weapon/storage/bag/P = O
 		var/plants_loaded = 0
 		for(var/obj/G in P.contents)
-			if(accept_check(G))
-				if(contents.len >= max_n_of_items)
-					user << "<span class='notice'>\The [src] is full.</span>"
-					return 1
-				else
-					P.remove_from_storage(G,src)
-					if(item_quants[G.name])
-						item_quants[G.name]++
-					else
-						item_quants[G.name] = 1
-					plants_loaded++
+			if(P.remove_from_storage(G, src) && accept_check(G))
+				plants_loaded++
+				stock_item(G)
+
 		if(plants_loaded)
-
-			user.visible_message("<span class='notice'>[user] loads \the [src] with \the [P].</span>", "<span class='notice'>You load \the [src] with \the [P].</span>")
+			user.visible_message("<span class='notice'>\The [user] loads \the [src] with \the [P].</span>", "<span class='notice'>You load \the [src] with \the [P].</span>")
 			if(P.contents.len > 0)
-				user << "<span class='notice'>Some items are refused.</span>"
-
-		nanomanager.update_uis(src)
+				user << "<span class='notice'>Some items were refused.</span>"
 
 	else
 		user << "<span class='notice'>\The [src] smartly refuses [O].</span>"
-		return 1
+	return 1
 
 /obj/machinery/smartfridge/secure/emag_act(var/remaining_charges, var/mob/user)
 	if(!emagged)
@@ -254,6 +238,20 @@
 		locked = -1
 		user << "You short out the product lock on [src]."
 		return 1
+
+/obj/machinery/smartfridge/proc/stock_item(var/obj/item/O)
+	for(var/datum/stored_items/I in item_records)
+		if(istype(O, I.item_path) && O.name == I.item_name)
+			stock(I, O)
+			return
+
+	var/datum/stored_items/I = new/datum/stored_items(src, O.type, O.name)
+	dd_insertObjectList(item_records, I)
+	stock(I, O)
+
+/obj/machinery/smartfridge/proc/stock(var/datum/stored_items/I, var/obj/item/O)
+	I.add_product(O)
+	nanomanager.update_uis(src)
 
 /obj/machinery/smartfridge/attack_ai(mob/user as mob)
 	attack_hand(user)
@@ -279,11 +277,11 @@
 	data["secure"] = is_secure
 
 	var/list/items[0]
-	for (var/i=1 to length(item_quants))
-		var/K = item_quants[i]
-		var/count = item_quants[K]
+	for (var/i=1 to length(item_records))
+		var/datum/stored_items/I = item_records[i]
+		var/count = I.get_amount()
 		if(count > 0)
-			items.Add(list(list("display_name" = html_encode(capitalize(K)), "vend" = i, "quantity" = count)))
+			items.Add(list(list("display_name" = html_encode(capitalize(I.item_name)), "vend" = i, "quantity" = count)))
 
 	if(items.len > 0)
 		data["contents"] = items
@@ -310,20 +308,15 @@
 	if(href_list["vend"])
 		var/index = text2num(href_list["vend"])
 		var/amount = text2num(href_list["amount"])
-		var/K = item_quants[index]
-		var/count = item_quants[K]
+		var/datum/stored_items/I = item_records[index]
+		var/count = I.get_amount()
 
 		// Sanity check, there are probably ways to press the button when it shouldn't be possible.
 		if(count > 0)
-			item_quants[K] = max(count - amount, 0)
-
-			var/i = amount
-			for(var/obj/O in contents)
-				if(O.name == K)
-					O.loc = loc
-					i--
-					if(i <= 0)
-						return 1
+			if((count - amount) < 0)
+				amount = count
+			for(var/i = 1 to amount)
+				I.get_product(get_turf(src))
 
 		return 1
 	return 0
@@ -334,17 +327,12 @@
 	if(!target)
 		return 0
 
-	for (var/O in item_quants)
-		if(item_quants[O] <= 0) //Try to use a record that actually has something to dump.
+	for(var/datum/stored_items/I in src.item_records)
+		throw_item = I.get_product(loc)
+		if (!throw_item)
 			continue
-
-		item_quants[O]--
-		for(var/obj/T in contents)
-			if(T.name == O)
-				T.loc = src.loc
-				throw_item = T
-				break
 		break
+
 	if(!throw_item)
 		return 0
 	spawn(0)
